requirements.txt
torch
torchvision
matplotlib
numpy
opencv-python
Pillow
streamlit

